backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6
jinja2>=3.1.2

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-decouple>=3.8

# Authentication & Security
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9

# Data Processing
pandas>=2.1.0
numpy>=1.25.0
openpyxl>=3.1.2

# Synthetic Data Generation
mimesis>=11.1.0

# LLM & AI
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.6.0
tiktoken>=0.5.2


# Vector Database & Embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
# torch>=2.0.0
# transformers>=4.41.0

# Payment Processing
stripe>=7.8.0

# HTTP Client
httpx>=0.25.0

# Email
email-validator>=2.1.0

# Task Queue & Caching
celery>=5.3.0
redis>=5.0.0
flower>=2.0.1  # Optional: Celery monitoring tool

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
scikit-learn>=1.3.0
nltk>=3.8.1

# Additional ML/AI utilities (installed with torch/transformers)
# tqdm>=4.66.0  # Progress bars (comes with transformers)
# regex>=2024.0  # Advanced regex (comes with transformers)

# Development (Optional - comment out for production)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0

